feat: replace custom binary protocol with protobuf - add game.proto (Handshake, PlayerMove, PlayerSync, PlayerSyncBatch, Envelope/oneof), generate pb/game.pb.go, migrate server/handler/state/service/ticker to proto.Marshal/Unmarshal, update PlayerState Chan→ChanField/uint8→uint32/int8→int32, Redis keys unchanged for NestJS compat, add make proto/proto-win, SETUP.md

+++ internal/game/player/trangthai.go
@@ -2,17 +2,21 @@ package player
 
 // Trangthai enum - PHẢI khớp với Java enum TrangThai.java
 // Thứ tự ordinal quan trọng vì lưu thành byte.
+//
+// THAY ĐỔI SO VỚI CUSTOM BINARY:
+//   - byte → uint32 để khớp với kiểu proto (proto không có uint8/byte)
+//   - Logic và giá trị string giữ nguyên hoàn toàn → NestJS/Redis không bị ảnh hưởng
 const (
-	TrangthaiDungYen  byte = iota // 0
-	TrangthaiDiChuyen             // 1
-	TrangthaiNhay                 // 2
-	TrangthaiRoi                  // 3
-	TrangthaiBayNgang             // 4
-	TrangthaiThu                  // 5
-	TrangthaiGong                 // 6
+	TrangthaiDungYen  uint32 = iota // 0
+	TrangthaiDiChuyen               // 1
+	TrangthaiNhay                   // 2
+	TrangthaiRoi                    // 3
+	TrangthaiBayNgang               // 4
+	TrangthaiThu                    // 5
+	TrangthaiGong                   // 6
 )
 
-var trangthaiToString = map[byte]string{
+var trangthaiToString = map[uint32]string{
 	TrangthaiDungYen:  "DUNG_YEN",
 	TrangthaiDiChuyen: "DI_CHUYEN",
 	TrangthaiNhay:     "NHAY",
@@ -22,7 +26,7 @@ var trangthaiToString = map[byte]string{
 	TrangthaiGong:     "GONG",
 }
 
-var stringToTrangthai = map[string]byte{
+var stringToTrangthai = map[string]uint32{
 	"DUNG_YEN":  TrangthaiDungYen,
 	"DI_CHUYEN": TrangthaiDiChuyen,
 	"NHAY":      TrangthaiNhay,
@@ -32,14 +36,14 @@ var stringToTrangthai = map[string]byte{
 	"GONG":      TrangthaiGong,
 }
 
-func TrangthaiToString(v byte) string {
+func TrangthaiToString(v uint32) string {
 	if s, ok := trangthaiToString[v]; ok {
 		return s
 	}
-	return "DUNG_YEN" // fallback
+	return "DUNG_YEN"
 }
 
-func StringToTrangthai(s string) byte {
+func StringToTrangthai(s string) uint32 {
 	if v, ok := stringToTrangthai[s]; ok {
 		return v
 	}
